requirements.txt
#Groq
langchain-groq
groq


# Gemini
langchain-google-genai
google-generativeai

#Groq
langchain-groq

langchain
langchain-community
streamlit
python-dotenv
PyPDF2
ipykernel
pandas
-e .